create dev

+++ src/main/java/com/mtech/ique/queueservice/model/entity/QueueInfo.java
@@ -5,9 +5,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-//import javax.persistence.Entity;
-//import javax.persistence.Id;
-
 @NoArgsConstructor
 @AllArgsConstructor
 @Setter
@@ -16,7 +13,7 @@ public class QueueInfo {
 
     private long queueId;
 
-    private int waitingSize;
+    private long waitingSize;
 
     private long estimateWaitingTime;
